Se cambia el tipo de dato aceptado en BD_AR.FOLIO_TELECARGA a un Entero de 32 ya que solo en test esta para 64

+++ WebApiSgsElavon/Model/BdAr.cs
@@ -371,7 +371,7 @@ namespace WebApiSgsElavon.Model
         [Column("ID_CALIFICA_INTENTO_4")]
         public int? IdCalificaIntento4 { get; set; }
         [Column("FOLIO_TELECARGA")]
-        public long? FolioTelecarga { get; set; }
+        public int? FolioTelecarga { get; set; }
         [Column("ID_TIPO_EQUIPO")]
         public int? IdTipoEquipo { get; set; }
         [Column("CORREO_EJECUTIVO")]
